+++ nodekit-rs/build.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+maturin develop --release
+cargo run --bin stub_gen
