+++ nodekit-rs/nodekit_rs.pyi
@@ -1,0 +1,12 @@
+# This file is automatically generated by pyo3_stub_gen
+# ruff: noqa: E501, F401
+
+import typing
+
+@typing.final
+class Frame:
+    r"""
+    A frame of audio/visual data.
+    """
+    ...
+
